Added Debian CVS build scripts. To build a snapshot Debian release, simply run, from within the top level of the glib directory:
[ben@gilgamesh:~/src/gtk-snap/glib]% debian/build                        1:16AM

This will build a Debian snapshot release, updating debian/changelog, and
place the resultant .debs in .. (~/src/gtk-snap/ in this example).
The version numbers are automatically updated, and look like:

YYYYMMDD.XX

where YYYY is the four-digit year (Y10K problem!) and MM is the
month (01-12) and DD is the day (01-31). XX is the build number;
it starts at 01 and debian/build increments it if you build from
CVS more than once in a day. If you're doing more than 99 CVS
builds in one day you need your head checked.

*NOTE*! The debian/build script I've written does not check in the
changes it has made to debian/changelog; that'd be scary and probably
generate too many log files all the time.

This really doesn't matter *too* much, since debian/changelog is
kind of irrelevant with CVS builds. Just know that the scant
information that is in there will not be updated via CVS.

Also, for obvious reasons, the debian/build script I've written
disables PGP signing of the resultant .changes and .dsc file.
Since these packages are not going into any archives, this will
not be a problem.

Of course, all this doesn't mean much to you if you don't have the
Debian dpkg-dev tools and debhelper installed, so don't worry if
you have no idea what I'm talking about. :)