OK, I think I've got all the .cvsignore stuff taken care of, though we'll want to fine-tune things as we go, of cours...
Original commit message from CVS:
OK, I think I've got all the .cvsignore stuff taken care of, though we'll
want to fine-tune things as we go, of course.  Most of them are the same,
with some exceptions for directories that produce executables (those are
listed by name after the standard ones and a newline for separation).