Changed the way state is dealt with when a child is added to the bin.  The note states that the COMPLETE state should...
Original commit message from CVS:
Changed the way state is dealt with when a child is added to the bin.  The
note states that the COMPLETE state should probably reflect nothing more
than whether or not there is a child in the bin, not whether or not all
children are COMPLETE.  I need to write out a few scenarios for complex
pipeline manipulations to figure out how all the states should interact.
The idea is to maintain the ability to dynamically recofigure the
pipeline.