Make sure the Gdk thread lock is not held inside the gstreamer pipeline.
Original commit message from CVS:
Make sure the Gdk thread lock is not held inside the gstreamer pipeline.
Fix the videosink to not hold the Gdk lock when emitting a signal.
All Gtk GUI apps using GStreamer should now handle the Gdk locks when
handling signals from the pipeline (as it should be)