Folling wtays suggestion, I was going to start using xmlDocDump, to avoid a symlink attack on the temporary registry ...
Original commit message from CVS:
Folling wtays suggestion, I was going to start using xmlDocDump, to
avoid a symlink attack on the temporary registry file.  Unfortunately,
xmlDocDump doesn't give any indication whether its successful, so I've
#ifdefed this out and left the original in place.  Since the tmp file
is in /etc/gstreamer, this should be okay for the moment, but I shall
ask the libxml people to add some way of getting the success value of
DocDump so we can use that in future.