Antoher way of dealing with EOS. This proposal does not use the recursion to propagate the EOS signal. This implies t...
Original commit message from CVS:
Antoher way of dealing with EOS. This proposal does not use the recursion
to propagate the EOS signal. This implies that an element cannot deny an
EOS signal anymore but since the signal is generated when a NULL buffer is
pushed, somebody did something wrong anyway.