Add a proposed makefile to use in place of Makefile.am.
Original commit message from CVS:
Add a proposed makefile to use in place of Makefile.am.
This depends on automake 1.4d or later, but is much neater, avoiding
horrible hacks to pass special flags in.

I put it here for safekeeping until automake 1.5 is released, at which
point we can reasonably require it.