Add checks for appropriate versions of automake.
Original commit message from CVS:
Add checks for appropriate versions of automake.
Add check that automake has been patched if it is a version which requires
a patch.
Please check that this doesn't cause problems: it's a bit of a hairy test,
so I might well have mucked it up and broken some people's build systems.
Hope not, though. ;-)