Remove gstreamer.html from generated files, since it was just a broken copy of one of the other files.
Original commit message from CVS:
Remove gstreamer.html from generated files, since it was just a broken copy
of one of the other files.
Sort out cleanfiles to some extent.
Note: the make dependencies in this file are completely broken, but it's not    clear how to fix them.  The docs will be built at make install / dist time,
as a result.