Ensure that the gstmarshal.[ch] files are generated in time by generating both files in the same rule: previously, on...
Original commit message from CVS:
Ensure that the gstmarshal.[ch] files are generated in time by generating
both files in the same rule: previously, only the .c would be generated in
time.  Don't put the files in dists (add a dist-hook to take them out),
and clean the files when running distclean.