Ensure that the marshaller code #includes glib-object.h, and hence the neccessary definitions to enable it to compile.
Original commit message from CVS:
Ensure that the marshaller code #includes glib-object.h, and hence the
neccessary definitions to enable it to compile.

I don't pretend to understand how the marshalling stuff works - I just
know that it clearly needs this #include, or something similar.  If there's
a more correct fix, someone else can do it.  Possibly this is a
glib-genmarshal bug, but I've not got the time to go into that.