this is a major update which adds a tree view of the pipeline and lets you set individual element args through a dyna...
Original commit message from CVS:
this is a major update which adds a tree view of the pipeline and lets you set individual element args through a dynamically generated gui.  If you haven't tried this tool yet then try it now. The next step will be to add gui generation for dynamic params as well. Please email feedback and bug reports to the list