quoting from the comment i put in gststaticautoplug.c: it is likely that the plugin is not loaded yet. thus when it l...
Original commit message from CVS:
quoting from the comment i put in gststaticautoplug.c:

it is likely that the plugin is not loaded yet. thus when it loads it
will replace the elementfactory that gst built from the cache, and the
GST_OBJECT_NAME will no longer be valid. thus we must g_strdup its name.

this might be an implementation problem, i don't know, if a program keeps
a reference to a cached factory after a factory has been added on plugin
initialization. i really don't know though.