please don't try to change what is going to be build in Makefile.am files everything is done using a USE_ conditional...
Original commit message from CVS:
please don't try to change what is going to be build in Makefile.am files
everything is done using a USE_ conditional, so set it to no in
configure.ac if you don't want it.  This one is a good example.

Don't know if qcam needs nasm though : is checking for nasm a check for
what hardware platform ? If so we need to change this.

In any case, let's not "fix" "problems" like this because this is how
cruft comes in.  Keep it in a local tree and bring it up so we can fix it
right ;)