added a test for pth and pthreads. in summary, it doesn't work because in __pthread_find_self() some funny stack poin...
Original commit message from CVS:
added a test for pth and pthreads. in summary, it doesn't work because
in __pthread_find_self() some funny stack pointer math is done that
will result in no matches for __pthread_find_self(). this is dereferenced in
__errno_location, causing a segfault on context switch because of the errno-swapping
on the part of pth. it's a mess and i don't know how to get around it. i'll
have to look into what ngpt does for looking up errno, as it has kernel threads as
well as user threads.