this script runs gst-inspect on all built elements and checks for warnings or bad exit codes.
Original commit message from CVS:
this script runs gst-inspect on all built elements and checks for warnings or bad exit codes.

It requires a gst-inspect to be in your PATH. more checking is probably needed for valid gst-inspect output.
The output format of gst-inspect will be changed slightly to make parsing the output possible.