Check to see if the registry is out of date, and if so try to rebuild it.
Original commit message from CVS:
Check to see if the registry is out of date, and if so try to rebuild it.

WIERD THING: trying to run gst-register using system() fails. Apparently child processes
can't link to libs parent processes use? It's wierd, and it could be a libtool issue.

That's probably just as well, it would be better just to use -register's code from within
core and not rely on another binary.